move participant-visits tests from tests/visits.py to tests/participants.py
